fix(lint): mechanical ruff batch — strict zips, subprocess checks, small cleanups
- zip(..., strict=True) on all 17 same-length-by-construction pairings
  (scoring, pareto/traces plotting, optimize callbacks, video select)
- explicit check=False where the return code is handled manually
  (ffmpeg wrapper, taskkill tree-kill)
- exception chaining in atomic-replace fallback (raise ... from)
- with-block for the per-slot SOFA log handle (closed right after Popen)
- justified noqa where the pattern is deliberate: Popen log handles that
  outlive the function (SIM115), module/process singletons (PLW0603),
  toy-map seed (S311), taskkill from PATH (S607)
- misc: collapsible if, implicit Optional, in-dict-keys, ambiguous names,
  redundant int(round()), lambda-to-def, unused loop var

116 tests green.

+++ src/sofaopt/sensitivity.py
@@ -136,7 +136,7 @@ def _build_samples(param: ParamSpec, n_samples: int) -> list[Any]:
     lo, hi = float(param.low), float(param.high)
     step = (hi - lo) / (n_samples - 1)
     if param.type == "int":
-        return sorted({int(round(lo + i * step)) for i in range(n_samples)})
+        return sorted({round(lo + i * step) for i in range(n_samples)})
     return [lo + i * step for i in range(n_samples)]
